% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{abs_energy}
\alias{abs_energy}
\title{Abs energy}
\usage{
abs_energy(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the absolute energy of the time series which is the sum over the squared values
}
\details{
\deqn{E = \sum_{i=1,\ldots, n} x_i^2}
}
