% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{absolute_sum_of_changes}
\alias{absolute_sum_of_changes}
\title{Absolute sum of changes}
\usage{
absolute_sum_of_changes(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the sum over the absolute value of consecutive changes in the series x
}
\details{
\deqn{\sum_{i=1, \ldots, n-1} \mid x_{i+1}- x_i \mid}
}
