% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{agg_autocorrelation}
\alias{agg_autocorrelation}
\title{Agg autocorrelation}
\usage{
agg_autocorrelation(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"f_agg": x, "maxlag", n} with x str, the name of a numpy function
(e.g. "mean", "var", "std", "median"), its the name of the aggregator function that is applied to the
autocorrelations. Further, n is an int and the maximal number of lags to consider.
(type:  list)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Calculates the value of an aggregation function \eqn{f_{agg}} (e.g. the variance or the mean) over the
autocorrelation \eqn{R(l)} for different lags. The autocorrelation \eqn{R(l)} for lag \eqn{l} is defined as
}
\details{
\deqn{R(l) = \frac{1}{(n-l)\sigma^{2}} \sum_{t=1}^{n-l}(X_{t}-\mu )(X_{t+l}-\mu)}

where \eqn{X_i} are the values of the time series, \eqn{n} its length. Finally, \eqn{\sigma^2} and
\eqn{\mu} are estimators for its variance and mean
(See \code{Estimation of the Autocorrelation function <http://en.wikipedia.org/wiki/Autocorrelation#Estimation>}_).

The \eqn{R(l)} for different lags \eqn{l} form a vector. This feature calculator applies the aggregation
function \eqn{f_{agg}} to this vector and returns

\deqn{f_{agg} \left( R(1), \ldots, R(m)\right) \quad \text{for} \quad m = max(n, maxlag).}

Here \eqn{maxlag} is the second parameter passed to this function.
}
