% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{agg_linear_trend}
\alias{agg_linear_trend}
\title{Agg linear trend}
\usage{
agg_linear_trend(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"attr": x, "chunk_len": l, "f_agg": f} with x, f an string and l an int
(type:  list)}
}
\value{
the different feature values
(type:  numeric vector)
}
\description{
Calculates a linear least-squares regression for values of the time series that were aggregated over chunks versus
the sequence from 0 up to the number of chunks minus one.
}
\details{
This feature assumes the signal to be uniformly sampled. It will not use the time stamps to fit the model.

The parameters attr controls which of the characteristics are returned. Possible extracted attributes are "pvalue",
"rvalue", "intercept", "slope", "stderr", see the documentation of linregress for more information.

The chunksize is regulated by "chunk_len". It specifies how many time series values are in each chunk.

Further, the aggregation function is controlled by "f_agg", which can use "max", "min" or , "mean", "median"
}
