% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{approximate_entropy}
\alias{approximate_entropy}
\title{Approximate entropy}
\usage{
approximate_entropy(x, m, r)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{m}{Length of compared run of data
(type:  int)}

\item{r}{Filtering level, must be positive
(type:  float)}
}
\value{
Approximate entropy
(type:  float)
}
\description{
Implements a vectorized Approximate entropy algorithm.
}
\details{
https://en.wikipedia.org/wiki/Approximate_entropy

For short time-series this method is highly dependent on the parameters,
but should be stable for N > 2000, see:

Yentes et al. (2012) -
\emph{The Appropriate Use of Approximate Entropy and Sample Entropy with Short Data Sets}

Other shortcomings and alternatives discussed in:

Richman & Moorman (2000) -
\emph{Physiological time-series analysis using approximate entropy and sample entropy}
}
