% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{augmented_dickey_fuller}
\alias{augmented_dickey_fuller}
\title{Augmented dickey fuller}
\usage{
augmented_dickey_fuller(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"attr": x} with x str, either "teststat", "pvalue" or "usedlag"
(type:  list)}
}
\value{
the value of this feature
(type:  float)
}
\description{
The Augmented Dickey-Fuller test is a hypothesis test which checks whether a unit root is present in a time
series sample. This feature calculator returns the value of the respective test statistic.
}
\details{
See the statsmodels implementation for references and more details.
}
