% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{autocorrelation}
\alias{autocorrelation}
\title{Autocorrelation}
\usage{
autocorrelation(x, lag)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{lag}{the lag
(type:  int)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Calculates the autocorrelation of the specified lag, according to the formula [1]
}
\details{
\deqn{\frac{1}{(n-l)\sigma^{2}} \sum_{t=1}^{n-l}(X_{t}-\mu )(X_{t+l}-\mu)}

where \eqn{n} is the length of the time series \eqn{X_i}, \eqn{\sigma^2} its variance and \eqn{\mu} its
mean. \code{l} denotes the lag.
}
\references{
[1] https://en.wikipedia.org/wiki/Autocorrelation#Estimation
}
