% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{binned_entropy}
\alias{binned_entropy}
\title{Binned entropy}
\usage{
binned_entropy(x, max_bins)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{max_bins}{the maximal number of bins
(type:  int)}
}
\value{
the value of this feature
(type:  float)
}
\description{
First bins the values of x into max_bins equidistant bins.
Then calculates the value of
}
\details{
\deqn{- \sum_{k=0}^{min(max\_bins, len(x))} p_k log(p_k) \cdot \mathbf{1}_{(p_k > 0)}}

where \eqn{p_k} is the percentage of samples in bin \eqn{k}.
}
