% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{c3}
\alias{c3}
\title{C3}
\usage{
c3(x, lag)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{lag}{the lag that should be used in the calculation of the feature
(type:  int)}
}
\value{
the value of this feature
(type:  float)
}
\description{
This function calculates the value of
}
\details{
\deqn{\frac{1}{n-2lag} \sum_{i=0}^{n-2lag} x_{i + 2 \cdot lag}^2 \cdot x_{i + lag} \cdot x_{i}}

which is

\deqn{\mathbb{E}\[L^2(X)^2 \cdot L(X) \cdot X\]}

where \eqn{\mathbb{E}} is the mean and \eqn{L} is the lag operator. It was proposed in [1] as a measure of
non linearity in the time series.
}
\references{
|  [1] Schreiber, T. and Schmitz, A. (1997).
|  Discrimination power of measures for nonlinearity in a time series
|  PHYSICAL REVIEW E, VOLUME 55, NUMBER 5
}
