% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{change_quantiles}
\alias{change_quantiles}
\title{Change quantiles}
\usage{
change_quantiles(x, ql, qh, isabs, f_agg)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{ql}{the lower quantile of the corridor
(type:  float)}

\item{qh}{the higher quantile of the corridor
(type:  float)}

\item{isabs}{should the absolute differences be taken?
(type:  bool)}

\item{f_agg}{the aggregator function that is applied to the differences in the bin
(type:  str, name of a numpy function (e.g. mean, var, std, median))}
}
\value{
the value of this feature
(type:  float)
}
\description{
First fixes a corridor given by the quantiles ql and qh of the distribution of x.
Then calculates the average, absolute value of consecutive changes of the series x inside this corridor.
}
\details{
Think about selecting a corridor on the
y-Axis and only calculating the mean of the absolute change of the time series inside this corridor.
}
