% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{cid_ce}
\alias{cid_ce}
\title{Cid ce}
\usage{
cid_ce(x, normalize)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{normalize}{should the time series be z-transformed?
(type:  bool)}
}
\value{
the value of this feature
(type:  float)
}
\description{
This function calculator is an estimate for a time series complexity [1] (A more complex time series has more peaks,
valleys etc.). It calculates the value of
}
\details{
\deqn{\sqrt{ \sum_{i=0}^{n-2lag} ( x_{i} - x_{i+1})^2 }}
}
\references{
|  [1] Batista, Gustavo EAPA, et al (2014).
|  CID: an efficient complexity-invariant distance for time series.
|  Data Mining and Knowledge Difscovery 28.3 (2014): 634-669.
}
