% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{count_above_mean}
\alias{count_above_mean}
\title{Count above mean}
\usage{
count_above_mean(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the number of values in x that are higher than the mean of x
}
