% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{cwt_coefficients}
\alias{cwt_coefficients}
\title{Cwt coefficients}
\usage{
cwt_coefficients(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"widths":x, "coeff": y, "w": z} with x array of int and y,z int
(type:  list)}
}
\value{
the different feature values
(type:  numeric vector)
}
\description{
Calculates a Continuous wavelet transform for the Ricker wavelet, also known as the "Mexican hat wavelet" which is
defined by
}
\details{
\deqn{\frac{2}{\sqrt{3a} \pi^{\frac{1}{4}}} (1 - \frac{x^2}{a^2}) exp(-\frac{x^2}{2a^2})}

where \eqn{a} is the width parameter of the wavelet function.

This feature calculator takes three different parameter: widths, coeff and w. The feature calculater takes all the
different widths arrays and then calculates the cwt one time for each different width array. Then the values for the
different coefficient for coeff and width w are returned. (For each dic in param one feature is returned)
}
