% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{energy_ratio_by_chunks}
\alias{energy_ratio_by_chunks}
\title{Energy ratio by chunks}
\usage{
energy_ratio_by_chunks(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"num_segments": N, "segment_focus": i} with N, i both ints}
}
\value{
the feature values
(type:  list of tuples (index, data))
}
\description{
Calculates the sum of squares of chunk i out of N chunks expressed as a ratio with the sum of squares over the whole
series.
}
\details{
Takes as input parameters the number num_segments of segments to divide the series into and segment_focus
which is the segment number (starting at zero) to return a feature on.

If the length of the time series is not a multiple of the number of segments, the remaining data points are
distributed on the bins starting from the first. For example, if your time series consists of 8 entries, the
first two bins will contain 3 and the last two values, e.g. \code{\[ 0.,  1.,  2.\], \[ 3.,  4.,  5.\]} and \code{\[ 6.,  7.\]}.

Note that the answer for \code{num_segments = 1} is a trivial "1" but we handle this scenario
in case somebody calls it. Sum of the ratios should be 1.0.
}
