% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{fft_aggregated}
\alias{fft_aggregated}
\title{Fft aggregated}
\usage{
fft_aggregated(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"aggtype": s} where s str and in ["centroid", "variance",
"skew", "kurtosis"]
(type:  list)}
}
\value{
the different feature values
(type:  numeric vector)
}
\description{
Returns the spectral centroid (mean), variance, skew, and kurtosis of the absolute fourier transform spectrum.
}
