% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{fft_coefficient}
\alias{fft_coefficient}
\title{Fft coefficient}
\usage{
fft_coefficient(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"coeff": x, "attr": s} with x int and x >= 0, s str and in ["real", "imag",
"abs", "angle"]
(type:  list)}
}
\value{
the different feature values
(type:  numeric vector)
}
\description{
Calculates the fourier coefficients of the one-dimensional discrete Fourier Transform for real input by fast
fourier transformation algorithm
}
\details{
\deqn{A_k =  \sum_{m=0}^{n-1} a_m \exp \left \{ -2 \pi i \frac{m k}{n} \right \}, \qquad k = 0,        \ldots , n-1.}

The resulting coefficients will be complex, this feature calculator can return the real part (attr=="real"),
the imaginary part (attr=="imag), the absolute value (attr=""abs) and the angle in degrees (attr=="angle).
}
