% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{friedrich_coefficients}
\alias{friedrich_coefficients}
\title{Friedrich coefficients}
\usage{
friedrich_coefficients(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"coeff": x} with x int and x >= 0
(type:  list)}

\item{c}{the time series name
(type:  str)}
}
\value{
the different feature values
(type:  numeric vector)
}
\description{
Coefficients of polynomial \eqn{h(x)}, which has been fitted to
the deterministic dynamics of Langevin model
}
\details{
\deqn{\dot{x}(t) = h(x(t)) + \mathcal{N}(0,R)}

as described by [1].

For short time-series this method is highly dependent on the parameters.
}
\references{
|  [1] Friedrich et al. (2000): Physics Letters A 271, p. 217-222
|  \emph{Extracting model equations from experimental data}
}
