% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{has_duplicate}
\alias{has_duplicate}
\title{Has duplicate}
\usage{
has_duplicate(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  bool)
}
\description{
Checks if any value in x occurs more than once
}
