% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{has_duplicate_max}
\alias{has_duplicate_max}
\title{Has duplicate max}
\usage{
has_duplicate_max(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  bool)
}
\description{
Checks if the maximum value of x is observed more than once
}
