% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tsfresh}
\alias{install_tsfresh}
\title{Install tsfresh and its dependencies}
\usage{
install_tsfresh(method = c("auto", "virtualenv", "conda"),
  conda = "auto", envname = "r-feasts-tsfresh",
  extra_packages = "pystan")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{envname}{Name of Python environment to install within}

\item{extra_packages}{Additional Python packages to install along with
tsfresh.}
}
\description{
Install tsfresh and its dependencies
}
