% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the kurtosis of x (calculated with the adjusted Fisher-Pearson standardized
moment coefficient G2).
}
