% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{large_standard_deviation}
\alias{large_standard_deviation}
\title{Large standard deviation}
\usage{
large_standard_deviation(x, r)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{r}{the percentage of the range to compare with
(type:  float)}
}
\value{
the value of this feature
(type:  bool)
}
\description{
Boolean variable denoting if the standard dev of x is higher
than 'r' times the range = difference between max and min of x.
Hence it checks if
}
\details{
\deqn{std(x) > r * (max(X)-min(X))}

According to a rule of the thumb, the standard deviation should be a forth of the range of the values.
}
