% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{last_location_of_maximum}
\alias{last_location_of_maximum}
\title{Last location of maximum}
\usage{
last_location_of_maximum(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the relative last location of the maximum value of x.
The position is calculated relatively to the length of x.
}
