% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{last_location_of_minimum}
\alias{last_location_of_minimum}
\title{Last location of minimum}
\usage{
last_location_of_minimum(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the last location of the minimal value of x.
The position is calculated relatively to the length of x.
}
