% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{length}
\alias{length}
\title{Length}
\usage{
length(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  int)
}
\description{
Returns the length of x
}
