% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{linear_trend}
\alias{linear_trend}
\title{Linear trend}
\usage{
linear_trend(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"attr": x} with x an string, the attribute name of the regression model
(type:  list)}
}
\value{
the different feature values
(type:  numeric vector)
}
\description{
Calculate a linear least-squares regression for the values of the time series versus the sequence from 0 to
length of the time series minus one.
This feature assumes the signal to be uniformly sampled. It will not use the time stamps to fit the model.
The parameters control which of the characteristics are returned.
}
\details{
Possible extracted attributes are "pvalue", "rvalue", "intercept", "slope", "stderr", see the documentation of
linregress for more information.
}
