% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{longest_strike_below_mean}
\alias{longest_strike_below_mean}
\title{Longest strike below mean}
\usage{
longest_strike_below_mean(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the length of the longest consecutive subsequence in x that is smaller than the mean of x
}
