% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{max_langevin_fixed_point}
\alias{max_langevin_fixed_point}
\title{Max langevin fixed point}
\usage{
max_langevin_fixed_point(x, m, r)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{m}{order of polynom to fit for estimating fixed points of dynamics
(type:  int)}

\item{r}{number of quantils to use for averaging
(type:  float)}
}
\value{
Largest fixed point of deterministic dynamics
(type:  float)
}
\description{
Largest fixed point of dynamics  :math:argmax_x {h(x)=0}` estimated from polynomial \eqn{h(x)},
which has been fitted to the deterministic dynamics of Langevin model
}
\details{
\deqn{\dot(x)(t) = h(x(t)) + R \mathcal(N)(0,1)}

as described by

Friedrich et al. (2000): Physics Letters A 271, p. 217-222
\emph{Extracting model equations from experimental data}

For short time-series this method is highly dependent on the parameters.
}
