% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{maximum}
\alias{maximum}
\title{Maximum}
\usage{
maximum(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Calculates the highest value of the time series x.
}
