% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{mean}
\alias{mean}
\title{Mean}
\usage{
mean(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the mean of x
}
