% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{mean_abs_change}
\alias{mean_abs_change}
\title{Mean abs change}
\usage{
mean_abs_change(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the mean over the absolute differences between subsequent time series values which is
}
\details{
\deqn{\frac{1}{n} \sum_{i=1,\ldots, n-1} | x_{i+1} - x_{i}|}
}
