% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{mean_change}
\alias{mean_change}
\title{Mean change}
\usage{
mean_change(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the mean over the differences between subsequent time series values which is
}
\details{
\deqn{\frac{1}{n} \sum_{i=1,\ldots, n-1}  x_{i+1} - x_{i}}
}
