% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{minimum}
\alias{minimum}
\title{Minimum}
\usage{
minimum(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Calculates the lowest value of the time series x.
}
