% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{number_crossing_m}
\alias{number_crossing_m}
\title{Number crossing m}
\usage{
number_crossing_m(x, m)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{m}{the threshold for the crossing
(type:  float)}
}
\value{
the value of this feature
(type:  int)
}
\description{
Calculates the number of crossings of x on m. A crossing is defined as two sequential values where the first value
is lower than m and the next is greater, or vice-versa. If you set m to zero, you will get the number of zero
crossings.
}
