% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{number_cwt_peaks}
\alias{number_cwt_peaks}
\title{Number cwt peaks}
\usage{
number_cwt_peaks(x, n)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{n}{maximum width to consider
(type:  int)}
}
\value{
the value of this feature
(type:  int)
}
\description{
This feature calculator searches for different peaks in x. To do so, x is smoothed by a ricker wavelet and for
widths ranging from 1 to n. This feature calculator returns the number of peaks that occur at enough width scales
and with sufficiently high Signal-to-Noise-Ratio (SNR)
}
