% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{number_peaks}
\alias{number_peaks}
\title{Number peaks}
\usage{
number_peaks(x, n)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{n}{the support of the peak
(type:  int)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Calculates the number of peaks of at least support n in the time series x. A peak of support n is defined as a
subsequence of x where a value occurs, which is bigger than its n neighbours to the left and to the right.
}
\details{
Hence in the sequence

x = [3, 0, 0, 4, 0, 0, 13]

4 is a peak of support 1 and 2 because in the subsequences

[0, 4, 0]
[0, 0, 4, 0, 0]

4 is still the highest value. Here, 4 is not a peak of support 3 because 13 is the 3th neighbour to the right of 4
and its bigger than 4.
}
