% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{percentage_of_reoccurring_datapoints_to_all_datapoints}
\alias{percentage_of_reoccurring_datapoints_to_all_datapoints}
\title{Percentage of reoccurring datapoints to all datapoints}
\usage{
percentage_of_reoccurring_datapoints_to_all_datapoints(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the percentage of unique values, that are present in the time series
more than once.
}
\details{
len(different values occurring more than once) / len(different values)

This means the percentage is normalized to the number of unique values,
in contrast to the percentage_of_reoccurring_values_to_all_values.
}
