% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{percentage_of_reoccurring_values_to_all_values}
\alias{percentage_of_reoccurring_values_to_all_values}
\title{Percentage of reoccurring values to all values}
\usage{
percentage_of_reoccurring_values_to_all_values(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the ratio of unique values, that are present in the time series
more than once.
}
\details{
of data points occurring more than once / # of all data points

This means the ratio is normalized to the number of data points in the time series,
in contrast to the percentage_of_reoccurring_datapoints_to_all_datapoints.
}
