% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{quantile}
\alias{quantile}
\title{Quantile}
\usage{
quantile(x, q)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{q}{the quantile to calculate
(type:  float)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Calculates the q quantile of x. This is the value of x greater than q% of the ordered values from x.
}
