% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{range_count}
\alias{range_count}
\title{Range count}
\usage{
range_count(x, min, max)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{min}{the inclusive lower bound of the range
(type:  int or float)}

\item{max}{the exclusive upper bound of the range
(type:  int or float)}
}
\value{
the count of values within the range
(type:  int)
}
\description{
Count observed values within the interval [min, max).
}
