% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ratio_beyond_r_sigma}
\alias{ratio_beyond_r_sigma}
\title{Ratio beyond r sigma}
\usage{
ratio_beyond_r_sigma(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  iterable)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Ratio of values that are more than r*std(x) (so r sigma) away from the mean of x.
}
