% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ratio_value_number_to_time_series_length}
\alias{ratio_value_number_to_time_series_length}
\title{Ratio value number to time series length}
\usage{
ratio_value_number_to_time_series_length(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns a factor which is 1 if all values in the time series occur only once,
and below one if this is not the case.
In principle, it just returns
}
\details{
unique values / # values
}
