% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sample_entropy}
\alias{sample_entropy}
\title{Sample entropy}
\usage{
sample_entropy(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Calculate and return sample entropy of x.
}
\references{
|  [1] http://en.wikipedia.org/wiki/Sample_Entropy
|  [2] https://www.ncbi.nlm.nih.gov/pubmed/10843903?dopt=Abstract
}
