% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{spkt_welch_density}
\alias{spkt_welch_density}
\title{Spkt welch density}
\usage{
spkt_welch_density(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{param}{contains dictionaries {"coeff": x} with x int
(type:  list)}
}
\value{
the different feature values
(type:  numeric vector)
}
\description{
This feature calculator estimates the cross power spectral density of the time series x at different frequencies.
To do so, the time series is first shifted from the time domain to the frequency domain.
}
\details{
The feature calculators returns the power spectrum of the different frequencies.
}
