% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Standard deviation}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the standard deviation of x
}
