% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sum_of_reoccurring_values}
\alias{sum_of_reoccurring_values}
\title{Sum of reoccurring values}
\usage{
sum_of_reoccurring_values(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  float)
}
\description{
Returns the sum of all values, that are present in the time series
more than once.
}
