% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sum_values}
\alias{sum_values}
\title{Sum values}
\usage{
sum_values(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  bool)
}
\description{
Calculates the sum over the time series values
}
