% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{symmetry_looking}
\alias{symmetry_looking}
\title{Symmetry looking}
\usage{
symmetry_looking(x, param)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{r}{the percentage of the range to compare with
(type:  float)}
}
\value{
the value of this feature
(type:  bool)
}
\description{
Boolean variable denoting if the distribution of x \emph{looks symmetric}. This is the case if
}
\details{
\deqn{| mean(X)-median(X)| < r * (max(X)-min(X))}
}
