% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{time_reversal_asymmetry_statistic}
\alias{time_reversal_asymmetry_statistic}
\title{Time reversal asymmetry statistic}
\usage{
time_reversal_asymmetry_statistic(x, lag)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{lag}{the lag that should be used in the calculation of the feature
(type:  int)}
}
\value{
the value of this feature
(type:  float)
}
\description{
This function calculates the value of
}
\details{
\deqn{\frac{1}{n-2lag} \sum_{i=0}^{n-2lag} x_{i + 2 \cdot lag}^2 \cdot x_{i + lag} - x_{i + lag} \cdot  x_{i}^2}

which is

\deqn{\mathbb{E}\[L^2(X)^2 \cdot L(X) - L(X) \cdot X^2\]}

where \eqn{\mathbb{E}} is the mean and \eqn{L} is the lag operator. It was proposed in [1] as a
promising feature to extract from time series.
}
\references{
|  [1] Fulcher, B.D., Jones, N.S. (2014).
|  Highly comparative feature-based time-series classification.
|  Knowledge and Data Engineering, IEEE Transactions on 26, 3026–3037.
}
