% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{tsfresh_features}
\alias{tsfresh_features}
\title{Extract tsfresh features}
\usage{
tsfresh_features(x, ...)
}
\arguments{
\item{x}{A vector (must be sorted in time order)}

\item{...}{Unused.}
}
\description{
Extract tsfresh features
}
