% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{value_count}
\alias{value_count}
\title{Value count}
\usage{
value_count(x, value)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}

\item{value}{the value to be counted
(type:  int or float)}
}
\value{
the count
(type:  int)
}
\description{
Count occurrences of \code{value} in time series x.
}
