% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{variance_larger_than_standard_deviation}
\alias{variance_larger_than_standard_deviation}
\title{Variance larger than standard deviation}
\usage{
variance_larger_than_standard_deviation(x)
}
\arguments{
\item{x}{the time series to calculate the feature of
(type:  numeric vector)}
}
\value{
the value of this feature
(type:  bool)
}
\description{
Boolean variable denoting if the variance of x is greater than its standard deviation. Is equal to variance of x
being larger than 1
}
